# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FrozenBot API
  description: |
    API for FrozenBot - Telegram bot for frozen food delivery system.

    This API serves three main clients:
    - **Telegram Bot**: For user interactions and order processing
    - **Admin Panel**: React-based admin interface for managing orders, products, and analytics
    - **WebApp**: Telegram WebApp for customer ordering interface

    ## Authentication

    The API uses multiple authentication methods:
    - **JWT Tokens**: For admin panel authentication
    - **Telegram WebApp Validation**: For WebApp users using initData
    - **Bot Token**: For internal bot operations

    ## Rate Limiting

    API endpoints are rate limited to prevent abuse:
    - Admin endpoints: 100 requests per minute
    - User endpoints: 60 requests per minute
    - Public endpoints: 30 requests per minute

  version: 1.0.0
  contact:
    name: FrozenBot API Support
    email: support@frozenbot.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.frozenbot.com
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Admin login
      description: Authenticate admin user and return JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh JWT token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  # Users
...
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

    TelegramWebApp:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: Telegram WebApp initData for user authentication
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
    ValidationError:
      type: object
      required:
        - error
        - message
        - details
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Validation failed"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string

    # Authentication
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          format: password
          example: "password123"
    LoginResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 1800
        user:
          $ref: '#/components/schemas/User'
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token
    TokenResponse:
      type: object
      required:
        - access_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
          description: New JWT access token
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 1800

    # User
    User:
      type: object
      required:
        - id
        - telegram_id
        - first_name
        - is_admin
        - is_active
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        telegram_id:
          type: integer
          format: int64
          example: 123456789
        username:
          type: string
          nullable: true
          example: "johndoe"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          nullable: true
          example: "Doe"
        phone:
          type: string
          nullable: true
          example: "+1234567890"
        is_admin:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
        full_name:
          type: string
          readOnly: true
          example: "John Doe"
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "bad_request"
            message: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Authentication required"
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: "validation_error"
            message: "Validation failed"
            details:
              - field: "price"
                message: "Price must be greater than 0"
                code: "value_error.number.not_gt"